+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.2] - 2021-03-31
+
+- Add executable path setting for Golang and Python
+- Fix Python parameters not working for version less than Python 3.6 - [issue #4](https://github.com/RobertOstermann/vscode-inline-parameters/issues/4)
+
 ## [1.1.1] - 2021-03-31
 
 - Fix regex for parsing Golang parameters - [issue #3](https://github.com/RobertOstermann/vscode-inline-parameters/issues/3)